pyproject.toml
[tool.poetry]
name = "goit-hw-02"
version = "0.1.0"
description = ""
authors = ["Roman <Sudorenkoroma@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
faker = "^24.2.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
